+++ src/iqm_client/iqm_client.py
@@ -95,6 +95,7 @@ import json
 import os
 import time
 import warnings
+from base64 import b64decode
 from datetime import datetime
 from enum import Enum
 from posixpath import join
@@ -106,10 +107,19 @@ from pydantic import BaseModel, Field
 
 DEFAULT_TIMEOUT_SECONDS = 900
 SECONDS_BETWEEN_CALLS = 1
+REFRESH_MARGIN_SECONDS = 5
+
+AUTH_CLIENT_ID = 'iqm_client'
+AUTH_REALM = 'cortex'
 
 
 class ClientConfigurationError(RuntimeError):
     """Wrong configuration provided.
+    """
+
+
+class ClientAuthenticationError(RuntimeError):
+    """Something went wrong with user authentication.
     """
 
 
@@ -219,17 +229,90 @@ class RunResult(BaseModel):
         return RunResult(status=RunStatus(input_copy.pop('status')), **input_copy)
 
 
-def _get_credentials(username: Optional[str], api_key: Optional[str]) -> Optional[tuple[str, str]]:
+class GrantType(str, Enum):
+    """
+    Type of token request.
+    """
+    PASSWORD = 'password'
+    REFRESH = 'refresh_token'
+
+
+class AuthRequest(BaseModel):
+    """Request sent to authentication server for access token and refresh token, or for terminating the session.
+
+    * Token request with grant type ``'password'`` starts a new session in the authentication server.
+      It uses fields ``client_id``, ``grant_type``, ``username`` and ``password``.
+    * Token request with grant type ``'refresh_token'`` is used for maintaining an existing session.
+      It uses field ``client_id``, ``grant_type``, ``refresh_token``.
+    * Logout request uses only fields ``client_id`` and ``refresh_token``.
+
+    """
+    client_id: str = Field(..., description='name of the client for all request types')
+    'name of the client for all request types'
+    grant_type: Optional[GrantType] = Field(
+        None,
+        description="type of token request, in ``{'password', 'refresh_token'}``"
+    )
+    "type of token request, in ``{'password', 'refresh_token'}``"
+    username: Optional[str] = Field(None, description="username for grant type ``'password'``")
+    "username for grant type ``'password'``"
+    password: Optional[str] = Field(None, description="password for grant type ``'password'``")
+    "password for grant type ``'password'``"
+    refresh_token: Optional[str] = Field(
+        None,
+        description="refresh token for grant type ``'refresh_token'`` and logout request")
+    "refresh token for grant type ``'refresh_token'`` and logout request"
+
+
+class Credentials(BaseModel):
+    """Credentials and tokens for maintaining a session with the authentication server.
+
+    * Fields ``auth:server_url``, ``username`` and ``password`` are provided by the user.
+    * Fields ``access_token`` and ``refresh_token`` are loaded from the authentication server and
+      refreshed periodically.
+    """
+    auth_server_url: str = Field(..., description='Base URL of the authentication server')
+    'Base URL of the authentication server'
+    username: str = Field(..., description='username for logging in to the server')
+    'username for logging in to the server'
+    password: str = Field(..., description='password for logging in to the server')
+    'password for logging in to the server'
+    access_token: Optional[str] = Field(None, description='current access token of the session')
+    'current access token of the session'
+    refresh_token: Optional[str] = Field(None, description='current refresh token of the session')
+    'current refresh token of the session'
+
+
+def _get_credentials(credentials: dict[str, str]) -> Optional[Credentials]:
     """Try to obtain credentials, first from arguments, then from environment variables.
 
+    Args:
+        credentials: dict of credentials provided as arguments
+
     Returns:
-        (username, API key), or ``None`` if credentials could not be found.
+        Credentials with token fields cleared, or None if ``auth_server_url`` was not set.
     """
-    uname = username or os.environ.get('IQM_SERVER_USERNAME')
-    key = api_key or os.environ.get('IQM_SERVER_API_KEY')
-    if uname and key:
-        return uname, key
-    return None
+    auth_server_url = credentials.get('auth_server_url') or os.environ.get('IQM_AUTH_SERVER')
+    username = credentials.get('username') or os.environ.get('IQM_AUTH_USERNAME')
+    password = credentials.get('password') or os.environ.get('IQM_AUTH_PASSWORD')
+    if not auth_server_url:
+        return None
+    if not username or not password:
+        raise ClientConfigurationError('Auth server URL is set but no username or password')
+    return Credentials(auth_server_url=auth_server_url, username=username, password=password)
+
+
+def _time_left_seconds(token: str) -> int:
+    """Check how much time is left until the token expires.
+
+    Returns:
+        Time left on token in seconds.
+    """
+    _, body, _ = token.split('.', 2)
+    # Add padding to adjust body length to a multiple of 4 chars as required by base64 decoding
+    body += '=' * (-len(body) % 4)
+    exp_time = int(json.loads(b64decode(body)).get('exp', '0'))
+    return max(0, exp_time - int(time.time()))
 
 
 class IQMClient:
@@ -238,23 +321,30 @@ class IQMClient:
     Args:
         url: Endpoint for accessing the server. Has to start with http or https.
         settings: Settings for the quantum computer, in IQM JSON format.
-        username: Username, if required by the IQM server.
-            This can also be set in the IQM_SERVER_USERNAME environment variable.
-        api_key: API key, if required by the IQM server.
-            This can also be set in the IQM_SERVER_API_KEY environment variable.
+
+    Keyword Args:
+        auth_server_url: Optional base URL of the authentication server.
+            This can also be set in the IQM_AUTH_SERVER environment variable.
+            If unset, requests will be sent unauthenticated.
+        username: Optional username to log in to authentication server.
+            This can also be set in the IQM_AUTH_USERNAME environment variable.
+            Username must be set if ``auth_server_url`` is set.
+        password: Optional password to log in to authentication server.
+            This can also be set in the IQM_AUTH_PASSWORD environment variable.
+            Password must be set if ``auth_server_url`` is set.
     """
     def __init__(
             self,
             url: str,
             settings: dict[str, Any],
-            username: Optional[str] = None,
-            api_key: Optional[str] = None
+            **credentials  # contains auth_server_url, username, password
     ):
         if not url.startswith(('http:', 'https:')):
             raise ClientConfigurationError(f'The URL schema has to be http or https. Incorrect schema in URL: {url}')
         self._base_url = url
         self._settings = settings
-        self._credentials = _get_credentials(username, api_key)
+        self._credentials = _get_credentials(credentials)
+        self._update_tokens()
 
     def submit_circuit(
             self,
@@ -274,6 +364,8 @@ class IQMClient:
             ID for the created task. This ID is needed to query the status and the execution results.
         """
 
+        bearer_token = self._get_bearer_token()
+
         data = RunRequest(
             qubit_mapping=qubit_mapping,
             circuit=circuit,
@@ -282,14 +374,16 @@ class IQMClient:
         )
 
         headers = {'Expect': '100-Continue'}
+        if bearer_token:
+            headers['Authorization'] = bearer_token
+
         result = requests.post(
             join(self._base_url, 'circuit/run'),
             json=data.dict(),
             headers=headers,
-            auth=self._credentials
         )
         result.raise_for_status()
-        return UUID(json.loads(result.text)['id'])
+        return UUID(result.json()['id'])
 
     def get_run(self, run_id: UUID) -> RunResult:
         """Query the status of the running task.
@@ -304,9 +398,13 @@ class IQMClient:
             HTTPException: http exceptions
             CircuitExecutionError: IQM server specific exceptions
         """
-        result = requests.get(join(self._base_url, 'circuit/run/', str(run_id)), auth=self._credentials)
+        bearer_token = self._get_bearer_token()
+        result = requests.get(
+            join(self._base_url, 'circuit/run/', str(run_id)),
+            headers=None if not bearer_token else {'Authorization': bearer_token}
+        )
         result.raise_for_status()
-        result = RunResult.from_dict(json.loads(result.text))
+        result = RunResult.from_dict(result.json())
         if result.warnings:
             for warning in result.warnings:
                 warnings.warn(warning)
@@ -334,3 +432,86 @@ class IQMClient:
                 return results
             time.sleep(SECONDS_BETWEEN_CALLS)
         raise APITimeoutError(f"The task didn't finish in {timeout_secs} seconds.")
+
+    def close(self) -> bool:
+        """Terminate session with authentication server.
+
+        Returns:
+            True iff session was successfully closed
+
+        Raises:
+            ClientAuthenticationError: if logout failed
+        """
+        if self._credentials is None:
+            return False
+
+        if not self._credentials.refresh_token:
+            return False
+
+        url = f'{self._credentials.auth_server_url}/realms/{AUTH_REALM}/protocol/openid-connect/logout'
+        data = AuthRequest(client_id=AUTH_CLIENT_ID, refresh_token=self._credentials.refresh_token)
+        result = requests.post(url, data=data.dict(exclude_none=True))
+        if result.status_code not in [200, 204]:
+            raise ClientAuthenticationError(f'Logout failed, {result.text}')
+        self._credentials.access_token = None
+        self._credentials.refresh_token = None
+        return True
+
+    def _get_bearer_token(self, retries: int = 1) -> Optional[str]:
+        """Make a bearer token for Authorization header. If access token is about to expire refresh it first.
+
+        Args:
+            retries: number of times to try updating the tokens
+
+        Returns:
+            Bearer token, i.e. string containing prefix 'Bearer ' and the access token, or None if access token
+            is not available.
+        """
+        if self._credentials is None or not self._credentials.access_token:
+            return None
+        if _time_left_seconds(self._credentials.access_token) > REFRESH_MARGIN_SECONDS:
+            # Access token is still valid, no need to refresh
+            return f'Bearer {self._credentials.access_token}'
+        if retries < 1:
+            return None
+        self._update_tokens()
+        return self._get_bearer_token(retries - 1)
+
+    def _update_tokens(self):
+        """Update access token and refresh token.
+
+        Uses refresh token to request new tokens from authentication server.  If the refresh token has expired or
+        is about to expire, starts a new session by requesting the tokens using username and password.
+
+        Updated tokens are stored in ``self._credentials``.
+
+        Raises:
+            ClientAuthenticationError: updating the tokens failed
+        """
+        if self._credentials is None:
+            return
+
+        refresh_token = self._credentials.refresh_token
+        if refresh_token and _time_left_seconds(refresh_token) > REFRESH_MARGIN_SECONDS:
+            # Update tokens using existing refresh_token
+            data = AuthRequest(
+                client_id=AUTH_CLIENT_ID,
+                grant_type=GrantType.REFRESH,
+                refresh_token=refresh_token
+            )
+        else:
+            # Update tokens using username and password
+            data = AuthRequest(
+                client_id=AUTH_CLIENT_ID,
+                grant_type=GrantType.PASSWORD,
+                username=self._credentials.username,
+                password=self._credentials.password
+            )
+
+        url = f'{self._credentials.auth_server_url}/realms/{AUTH_REALM}/protocol/openid-connect/token'
+        result = requests.post(url, data=data.dict(exclude_none=True))
+        if result.status_code != 200:
+            raise ClientAuthenticationError(f'Failed to update tokens, {result.text}')
+        tokens = result.json()
+        self._credentials.access_token = tokens.get('access_token')
+        self._credentials.refresh_token = tokens.get('refresh_token')